requirements.txt
natsort
Pillow
chardet
xpinyin